A lot of stuff

+++ osa4/blogilista/controllers/blogs.js
@@ -1,33 +1,57 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1})
+
     response.json(blogs)
   })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
   try {
     const body = request.body
+    const user = request.user
+
+    if (!user) {
+      return response.status(400).json({ error: 'userId missing or not valid' })
+    }
 
     const blog = new Blog({
+      url: body.url,
       title: body.title,
       author: body.author,
-      url: body.url,
+      user: user._id,
       likes: body.likes
     })
 
     const savedBlog = await blog.save()
+
+    const freshUser = await User.findById(user._id)
+    freshUser.blogs = freshUser.blogs.concat(savedBlog._id)
+    await freshUser.save()
+
     response.status(201).json(savedBlog)
+
   } catch (error) {
     next(error)
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
   try {
+    const blog = await Blog.findById(request.params.id)
+    const user = request.user
+
+    if (blog.user.toString() !== user.id) {
+      return response.status(403).json({ error: 'only the creator can delete this blog' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
+
   } catch (error) {
     next(error)
   }
@@ -46,6 +70,7 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
     const updatedBlog = await blog.save()
     response.json(updatedBlog)
+    
   } catch (error) {
     next(error)
   }
